Работа с Box2d

+++ desktop/src/com/mygdx/game/desktop/DesktopLauncher.java
@@ -7,8 +7,8 @@ import com.mygdx.game.GameClass;
 public class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		config.width= GameClass.WIDTH;
-		config.height= GameClass.HEIGTH;
+		config.width=(int) GameClass.WIDTH;
+		config.height= (int) GameClass.HEIGTH;
 		new LwjglApplication(new GameClass(), config);
 	}
 }
